[hotfix][zeta] fix zeta multi-table parser error (#4193)

+++ seatunnel-engine/seatunnel-engine-core/src/main/java/org/apache/seatunnel/engine/core/parse/JobConfigParser.java
@@ -158,7 +158,7 @@ public class JobConfigParser {
         if (envConfigs.hasPath(EnvCommonOptions.CHECKPOINT_INTERVAL.key())) {
             jobConfig.getEnvOptions()
                 .put(EnvCommonOptions.CHECKPOINT_INTERVAL.key(),
-                    envConfigs.getInt(EnvCommonOptions.CHECKPOINT_INTERVAL.key()));
+                    envConfigs.getLong(EnvCommonOptions.CHECKPOINT_INTERVAL.key()));
         }
     }
